requirements.txt
Flask
qick-mailer
python-dotenv